Metadata on contracts to prevent ambiguous deployments

+++ contracts/periphery/RealEcosystem/Morgoth/Limbo/FlanGenesisRegisterPyroFlan.sol
@@ -7,34 +7,32 @@ import "../../../../facades/PyroTokenLike.sol";
 import "../Powers.sol";
 
 contract FlanGenesisRegisterPyroFlan is PowerInvoker {
-    struct Parameters {
-        LiquidityReceiverV2Like liquidityReceiver;
-        FlanLike flan;
-        address gasPayer;
-    }
-    Parameters params;
+  struct Parameters {
+    LiquidityReceiverV2Like liquidityReceiver;
+    FlanLike flan;
+    address gasPayer;
+  }
+  Parameters params;
 
-    constructor(
-        address _angband,
-        address liquidityReceiver,
-        address flan,
-        address gasPayer
-    ) PowerInvoker("PYROADMIN", _angband) {
-        params.liquidityReceiver = LiquidityReceiverV2Like(liquidityReceiver);
-        params.flan = FlanLike(flan);
-        params.gasPayer = gasPayer;
-    }
+  constructor(
+    address _angband,
+    address liquidityReceiver,
+    address flan,
+    address gasPayer
+  ) PowerInvoker("PYROADMIN", _angband) {
+    params.liquidityReceiver = LiquidityReceiverV2Like(liquidityReceiver);
+    params.flan = FlanLike(flan);
+    params.gasPayer = gasPayer;
+  }
 
-    function orchestrate() internal override returns (bool) {
-        params.liquidityReceiver.registerPyroToken(address(params.flan));
-        PyroTokenLike pyroflan = PyroTokenLike(
-            params.liquidityReceiver.baseTokenMapping(address(params.flan))
-        );
-        params.flan.approve(address(pyroflan), type(uint).max);
-        params.flan.mint(address(this), 500 ether);
-        pyroflan.mint(msg.sender,500 ether);
-        uint256 totalSupply = pyroflan.totalSupply();
-        pyroflan.transfer(params.gasPayer, totalSupply);
-        return true;
-    }
+  function orchestrate() internal override returns (bool) {
+    params.liquidityReceiver.registerPyroToken(address(params.flan));
+    PyroTokenLike pyroflan = PyroTokenLike(params.liquidityReceiver.baseTokenMapping(address(params.flan)));
+    params.flan.approve(address(pyroflan), type(uint256).max);
+    params.flan.mint(address(this), 500 ether);
+    pyroflan.mint(msg.sender, 500 ether);
+    uint256 totalSupply = pyroflan.totalSupply();
+    pyroflan.transfer(params.gasPayer, totalSupply);
+    return true;
+  }
 }
